Paas metrics move pingdumb into pkg
