+++ tools/metrics/pkg/pingdumb/pingdumb_test.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/alphagov/paas-cf/tools/metrics/fakes"
-	. "github.com/alphagov/paas-cf/tools/metrics/pingdumb"
+	. "github.com/alphagov/paas-cf/tools/metrics/pkg/pingdumb"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
